+++ ffw/products/templates/products/product.html
@@ -130,5 +130,4 @@
 {% endblock content %}
 
 {% block js %}
-    <script src="{{ STATIC_URL }}jsdev/product.js"> </script>
 {% endblock js %}
