+++ static/jsdev/product.js
@@ -1,9 +1,2 @@
-$(document).ready(function() {
 
-    cart.fulling();
 
-    $('button#buy').on('click', function() {
-        var product_pk = this.value;
-        cart.cart_change(product_pk,'add');
-    });
-});
